Dynamic Model

Quantifies the chill accumulation with dynamic equation (two-steps).

dynamic_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill for each temperature (TRUE is default).

Value

The function returns values the chill for each temperature of vector (Total = FALSE), or returns the chill accumulation (Total = TRUE).

Details

The model is based on dynamic accumulation, by means of the relationship between temperatures. The dynamic model assumes that the accumulated chill-units are not annulled by high temperatures.

Note

Code adapted from the function Dynamic_Model, of the chillR Package

References

FISHMAN, Svetlana, EREZ, A. & COUVILLON, G. A. (1987). The Temperature Dependence of Dormancy Breaking in Plants: Computer Simulation of Processes Studied Under Controlled Temperatures. J. Theor. Biol.

LUEDELING, Eike (2018). chillR: Statistical Methods for Phenology Analysis in Temperate Fruit Trees. R package version 0.70.12. https://CRAN.R-project.org/package=chillR

Examples


x <- rnorm(500, 10, 5)
dynamic_model(x)

#> [1] 8.196734

dynamic_model(x, FALSE)

#>   [1] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>   [6] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [11] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [16] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [21] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [26] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [31] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [36] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [41] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [46] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [51] 0.0000000000 0.0000000000 0.0000000000 1.0072060687 0.0000000000
#>  [56] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [61] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [66] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [71] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [76] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [81] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [86] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [91] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>  [96] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [101] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0001293652
#> [106] 1.0404826518 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [111] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [116] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [121] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [126] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [131] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [136] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [141] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [146] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [151] 0.0000000000 0.0000000000 0.0070293207 1.0087153272 0.0000000000
#> [156] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [161] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [166] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [171] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [176] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [181] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [186] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [191] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [196] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [201] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [206] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [211] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [216] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [221] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [226] 0.0000000000 0.0000000000 0.0000000000 0.9183990611 0.0000000000
#> [231] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [236] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [241] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [246] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [251] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [256] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [261] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [266] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [271] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [276] 0.0000000000 0.2054873070 0.0000000000 0.0000000000 0.0000000000
#> [281] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [286] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 1.0051668102
#> [291] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [296] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [301] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [306] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [311] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [316] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [321] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [326] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [331] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [336] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [341] 0.9846520845 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [346] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [351] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [356] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [361] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [366] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [371] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [376] 0.0000000000 0.0000000000 1.0063422802 0.0000000000 0.0000000000
#> [381] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [386] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [391] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [396] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [401] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [406] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [411] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [416] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [421] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [426] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 1.0131232898
#> [431] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [436] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [441] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [446] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [451] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [456] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [461] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [466] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [471] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [476] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [481] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [486] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [491] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [496] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000