GDH Model - Anderson - for heat accumulation

Quantifies the Growing Degree Hours at between the base and optimum temperatures, by Anderson et al. (1986).

gdha_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of GDH for each temperature (TRUE is default).

Value

The function returns values the GDH for each temperature of vector.

Details

The GDH Model is based on base and optimum temperature. The base temperature is 4°C. The optimum and critical temperature are 25°C and 36°C, respectively.

References

Anderson, J. L. et al. 1986. Validation of chill unit and flower bud phenology models for "Montmorency" sour cherry. Acta Horticulturae - Modelling in Fruit Research.

Examples


x <- rnorm(500,8,3)
gdha_model(x)

#> [1] 1268.74

gdha_model(x, FALSE)

#>   [1]  0.00000000000  2.81475890546  4.32003816064  0.46272666066  1.42018808861
#>   [6]  7.43778367895  1.51454294477  1.23132923468  4.10734650449  0.76058164858
#>  [11]  1.07440798491  3.39443711573  1.18122267790  3.84257014345  0.10767839298
#>  [16]  4.97590111614  0.48879513830  3.14067815374  1.41281187980  6.38186060757
#>  [21]  3.07126750645  1.69882767771  2.38451526012  0.92957928807  0.81030360279
#>  [26]  1.47160652594  0.00006999222  4.28032833703  6.77170752313  0.01339213565
#>  [31]  0.61345345059  0.02206728387  0.10881203585  0.84644410367  2.84273241800
#>  [36]  2.36253631961  6.46364097436  1.29839185787  3.68758706809  0.00000000000
#>  [41]  0.79531802142  4.25101165337  4.84333150405  0.03343824375  1.14637806842
#>  [46]  1.81585650623  7.47281334150  2.72662629247  0.79453696240  2.67635196539
#>  [51]  2.01726265139  0.00000000000  2.36783521459  1.00110621624  0.00000000000
#>  [56]  0.00000000000  2.21848940183  1.28459051681  3.68042938143  2.51139694083
#>  [61]  0.96108482623  2.58617772254  3.89779372665  0.87298377034  0.00000000000
#>  [66]  3.67824585235  0.41388812156  5.19576872596  1.91130448380  0.73265254338
#>  [71]  8.19378143349  0.08402076069  1.89954699555  3.45159388564  6.39109798177
#>  [76]  0.58261424458  0.85026681778  1.76398892835  2.02811218779  0.00000000000
#>  [81]  5.34782559514  3.79461442114  0.00000000000  1.94458864808  2.90937465228
#>  [86]  1.37905073326 10.19378962755  1.23355576498  1.85571700373  0.00000000000
#>  [91]  2.61705859606  0.52945740544  0.09853931374  4.18550577867  0.92227674097
#>  [96]  2.88198710036  0.35147649752  2.00538539457  2.56395026625  1.59031926752
#> [101]  0.38194009831  0.91576048642  0.00000000000  1.54578752429  2.02651447755
#> [106]  3.18792055847  2.97832262225  1.19858752117  7.51101960783  3.53742861955
#> [111]  2.79196551026  2.30918277123  3.10869096797  0.28424211671  3.89709976024
#> [116]  4.92689705145  2.98739007431  5.07439021993  1.39895158610  0.94653579721
#> [121]  2.22520102790  0.39748745302  2.40793503381  0.97141150762  4.41156795609
#> [126]  1.08606322423  0.23688300874  5.60179025798  0.31659722735  0.01335044174
#> [131]  0.08082411103  0.11947033819  0.45827271407  3.94820662321 10.82703880823
#> [136]  0.56082693332  1.12677270455  0.25530419100  9.86437114750  3.11783592229
#> [141]  4.49854957012  1.38880115846  1.07478134696  0.73792196774  2.36209224436
#> [146]  0.05569014942  1.55200910158  7.22094814121  1.09333490988  4.29214926800
#> [151]  0.00000000000  3.69385845641  0.00000000000  1.60336785065  1.91110796490
#> [156]  2.66454214898  4.49051140551  2.78033034106  0.00000000000  3.44675589278
#> [161]  0.29597459209  1.24411219684  4.67114623226  0.77927245051  2.14978654676
#> [166]  6.29008608426  0.43677931195  0.00000000000  1.71214262199  1.30810125332
#> [171]  0.00000000000 10.82562905727  0.52589795569  0.02892258778  2.24065962377
#> [176]  4.69457042584  0.08706003146  2.30722923576  1.09428472546  4.86061108024
#> [181]  5.18991469331  8.05873067706  4.51225405802  4.55317399045  1.01453808384
#> [186]  3.05701028562  0.49961595138  5.01534345564  3.25381164924  0.47283260108
#> [191]  1.08915785171  4.43513115304  0.18464617802  0.78635473770  4.53204947260
#> [196]  7.49993176018  2.58868355344  3.77949280121  3.25889147758  3.23185346305
#> [201]  1.42209975479  1.73087947179  0.60276023606  4.63201284086  0.59166669537
#> [206]  0.72177304753  5.61345095724  0.92132760221  0.47514395237  1.44858473771
#> [211]  0.22089050576  8.31232529383  2.00954370864  3.47361822248  1.61537887696
#> [216] 11.81653548587  2.96699850706  0.08348320690  0.00000000000  0.51757396442
#> [221]  0.15211715107  1.61780706100  5.50790943811  6.74178639044  2.49015813080
#> [226]  2.80476802170  0.33998715573  4.33862076269  1.29496936436  2.96529674858
#> [231]  0.65444638488  0.00000000000  2.94782346447  0.41350373611  0.58336972669
#> [236]  2.09884185478  6.09508639229  3.92825108512  0.52354368149  0.67543649325
#> [241]  4.75432333484  1.69360432600 11.03567760237  6.86861780089  0.60441175382
#> [246]  0.00000000000  0.00826731900  4.84764608178  9.66567356477  0.00000000000
#> [251]  1.76354702629  2.40208909056  1.94360252719  3.80861252639  0.15723137337
#> [256]  0.04682421272  1.77090368374  6.19353013413  2.05095566349  0.00000000000
#> [261]  5.48401045594  3.75723401106  3.67357223234  1.65197623877  0.55127925210
#> [266]  5.03643556441  2.99749822001  0.93633972700  0.54953981136  0.93793700173
#> [271]  0.19664754601  4.74029978658  3.57446450814  0.00000000000  0.28596919044
#> [276]  2.50663633832 12.09290462915  0.13130472033  7.70415289850  0.34704394754
#> [281]  3.33695189906  1.05250499037  0.84209792904  7.49400450916  1.83426771820
#> [286]  0.90329299946  0.00000000000  0.15826001358  4.87287507796  0.00000000000
#> [291]  0.15357564165  5.77591569027  2.21268815546  2.09955881988  0.01753902604
#> [296]  1.60908262189  0.30262310021  5.66963217994  2.12243034330  0.02156413566
#> [301]  2.63398469780  4.36927582628  1.28122700315  1.47580663823  2.14184281834
#> [306]  7.55522400555  0.48173464805  4.68302659276  0.02242877179  3.49589286723
#> [311]  1.18101029728  4.29591558189  0.09625307432  5.35207322554  1.45758659194
#> [316]  2.12474168135  0.00000000000  0.00000000000  3.71098059302  1.01291334448
#> [321]  1.14054022291  0.49326172853  0.00394987181  1.38675657392  0.00000000000
#> [326]  3.63864125554  1.99461442916  8.66438218986  6.88579738063  7.99979354656
#> [331]  0.29005207305  2.08607972447  4.91422790028  3.26832487401  0.91690706559
#> [336]  3.11792445469  4.65383987421  1.39796221345  1.82239256966  4.69190687476
#> [341]  3.85682980257  2.11807420125  0.00000000000  3.20478492095  6.39672654919
#> [346]  0.00000000000  2.49716696853  5.47981800117  3.12556457431  0.00000000000
#> [351] 12.68879796252  1.74355384784  0.00000000000  8.95131696273  3.35312901820
#> [356]  2.26024455954  3.55564227299  4.02893960372  0.00000000000  4.76153053721
#> [361]  7.23072224528  1.54463649938  6.63004701807  5.83420349873  0.17124679557
#> [366]  1.34157937898  4.94298315192  1.14349457389  3.86952462928  6.10733961146
#> [371]  0.12699709829  6.79438973119  0.50838272401  0.00687368122  0.60204770869
#> [376]  0.00000000000  2.58081024940  0.00000000000  4.35566008725  0.51849181774
#> [381]  0.00000000000  5.43749647658  4.26997077661  4.44893588092  0.00000000000
#> [386]  0.03650619067  3.13869039731  5.46858924814  5.10177946831  4.05429425518
#> [391]  1.26104189663  0.16687205914  0.19556272857  1.74340686333  0.22748353243
#> [396] 10.59952811108  7.57902720965  0.00000000000  4.91050474306  0.90444053283
#> [401]  0.00618132195  6.01964157709  5.28502340876  0.07539705429  0.06767784458
#> [406]  0.00000000000  0.00000000000  0.00167197352  0.00000000000  5.77062880724
#> [411]  1.83519847024  2.27904331749  0.00000000000  5.93655299566  0.00392347263
#> [416]  0.05093272318  0.00073678983  1.38901598898  0.44372730707  0.53117806395
#> [421]  1.43395865388  0.06436487994  3.53347380925  6.93489102264  0.00000000000
#> [426] 10.03323558192  0.00000000000  4.69138136109  0.94718981339  9.00170743088
#> [431]  5.40441909366  0.00000000000  0.00000000000  1.62536127199  0.15140180346
#> [436]  4.10929894711  8.67348757271  4.71419599846  2.25078666938  1.70323820785
#> [441]  0.00000000000  0.06260723127  1.62738694071  0.00079978624  1.66230961205
#> [446]  2.83198828917  0.64394870709  4.90725918983  1.59870531802  1.75716319903
#> [451]  0.00659488104  9.15342213604  1.46793254575  1.55132414735  4.16882034738
#> [456]  3.11597949672  0.00000000000  4.35913454347  0.03020228748  1.84806579075
#> [461]  0.80514577224  0.11539454987  3.11347218625  2.39729603361  2.33477994593
#> [466]  3.58819180005  1.15660571759  0.00000000000  6.51467618993  0.00000000000
#> [471]  1.55866614744  2.69624233702  0.90203050454  2.81950736613  0.00000000000
#> [476]  1.70291442485  1.90948627714  2.63901655432  4.46833964882  1.30895756725
#> [481]  1.03289667597  0.80291814115  5.34931673310  7.21379716812  1.27257639976
#> [486]  2.66574093570  2.56988033600  1.85083821580  1.36460207052  8.10650595512
#> [491]  3.14295126417  4.63969826629  5.23483520132  5.52764009329  1.68539005048
#> [496]  5.22115808624  0.72694048829  0.04972778983  0.13585299150 13.38746320193