GDH Model - Richardson - for heat accumulation

Quantifies the Growing Degree Hours at between the base and optimum temperatures.

gdhr_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of GDH for each temperature (TRUE is default).

Value

The function returns values the GDH for each temperature of vector.

Details

The GDH model is based on the subtraction of the base temperature of each hourly temperature between 4.5°C and 25°C. The temperature of the base is 4.5°C. Therefore, accumulation under temperatures below 4.5°C and above 25°C is zero.

References

RICHARDSON, E.A. et al. 1975. Pheno-climatographyof spring 249 peach bud development. HortScience.

Examples


x <- rnorm(500,8,3)
gdhr_model(x)

#> [1] 1895.762

gdhr_model(x, FALSE)

#>   [1]  1.18005959 13.94161925  1.64773408  5.06234733  5.27762076  0.79895945
#>   [7]  7.92284396  6.75539781  0.00000000  0.00000000  0.00000000  5.43791285
#>  [13]  2.21860163  7.06948444  3.08042257  4.06818793  3.67275322  1.73826162
#>  [19]  2.77575626  5.79140117  1.46570980  6.01098242  3.62581840  6.43792880
#>  [25]  7.98828696  7.64350147  4.03156055  1.68632334  3.48076234  0.40724719
#>  [31]  1.71014210  7.19001600  0.00000000  5.13962385  6.53436602  7.60152381
#>  [37]  5.62351391  3.13245244  1.10177260  2.96191999  3.82835857  4.19277966
#>  [43]  0.00000000  5.36157743  1.57295444  5.48707344  4.23765283  0.95631894
#>  [49]  2.19569777  6.14952785  5.25895604  2.72992388  7.16040711 10.47766390
#>  [55]  0.00000000  3.50081440  6.92753471  2.83417959  2.87505622  5.74976814
#>  [61]  3.85281685  3.79706361  7.58506541  0.98303528  0.30518834  9.83950405
#>  [67]  5.03200156  7.10078910  6.63638017  5.66133082  5.00890522  6.36752364
#>  [73]  4.97638417  3.03215159  2.20081347  0.00000000  2.62896988  4.52069597
#>  [79]  4.20365617  3.53547181  5.31419896  5.71252929  0.80568954  4.01546407
#>  [85]  6.23970437  0.18453964  4.68035043  0.00000000  5.71841485  5.41511281
#>  [91]  0.00000000  0.00000000  9.70622168  0.00000000  0.17074880  3.66469016
#>  [97]  0.91223783  6.29232647  5.21351120  7.20449578  4.88600938  1.29737542
#> [103]  7.47008187  4.87188431  1.29055774  9.93083849  2.48552190  3.28708880
#> [109]  3.23536994  4.56527215  0.00000000  7.86800151  3.49860317  3.14712903
#> [115]  4.05244464  0.91639783  5.83793254  2.13780749  0.00000000  3.91249437
#> [121]  1.54050425  5.66010529  5.92530928  6.38735639  3.80941422  5.55908045
#> [127]  6.78559680  4.04567665  2.90937375  0.95581585  0.00000000  0.00000000
#> [133]  3.82420168  0.00000000  3.13074423  3.94246364  0.00000000  3.18203575
#> [139]  5.00954573 11.03804135  2.17585595  0.00000000  1.46555310  5.68654180
#> [145]  0.00000000  1.73518407  6.82561739  1.99286004  4.54189606  8.74444038
#> [151]  4.88856602  5.80360912  4.85304870  2.44226833  4.59936346  2.03562538
#> [157]  6.12017004  8.89634187  6.65238564 10.02569192  2.94931141  1.43962099
#> [163]  3.57587655  5.83109367  7.57413820  4.43583627  6.47359305  2.35430406
#> [169]  0.85921545  3.56920496  5.62519039  7.91321350  7.76955532  1.30472824
#> [175]  0.00000000  0.96032832  8.42443473  0.00000000  8.04490046  7.53339586
#> [181]  4.52665399  2.52598856  0.64723622  3.86778897 10.12879264  7.41478411
#> [187]  1.92345780  6.95490777  5.97850069  6.96873450  4.73291808  3.33185407
#> [193]  5.84053140  0.00000000  2.67905154  5.31115319  0.00000000  6.36676589
#> [199]  3.53856061  4.37430828  0.00000000  0.00000000  5.87153991  3.33213869
#> [205] 11.85396961  4.68623511 11.59388726  6.66983040  8.84560794  7.45606412
#> [211]  1.32775054  2.71012650  2.45422774  2.86024430  5.74599940  0.00000000
#> [217]  2.88286942  3.06944985  7.43287159  5.79371978  5.19340572  2.97403584
#> [223]  3.49227594  5.26624981  4.02595147  7.69076891  5.97624994  5.16419153
#> [229]  5.13613395  1.89084649  0.00000000  1.36297342  3.21630063  4.22904791
#> [235]  6.56133832  6.26978597  4.28953982  1.43724144  4.16258354  1.83768867
#> [241]  5.15553916  3.01450104  7.55421060  3.48400057  0.00000000  6.46942808
#> [247]  0.00000000  3.58223169  2.32698430 10.70543053  5.16258800  5.18232784
#> [253]  0.00000000  4.26959281  4.84253895  6.22746365  1.52716462  4.26434388
#> [259]  0.00000000  0.04502587  3.77706856  1.92167289  0.67615251  0.01884464
#> [265]  7.88965037  0.00000000  7.36736692  5.06595024  3.61594074  3.51644270
#> [271]  9.65288813  5.25201046  1.34107887  4.45187592  4.00014696 11.00803560
#> [277]  4.70122613  6.71633596  3.66500289  3.31590842  0.27066923  5.01291288
#> [283]  5.10867365  0.00000000  4.18251892  0.00000000  1.04440657  0.00000000
#> [289]  0.25252003  4.11741308  0.25645304  2.93362234  4.43826936  3.72415100
#> [295]  1.68696544  3.63573995  8.44561891  1.23295105  2.63546194  1.30954848
#> [301]  5.07244556  1.72489841  0.00000000  7.10105940  0.77578602  0.00000000
#> [307]  4.72413486  6.11702846  0.00000000  4.11835281  0.30342200  4.83283733
#> [313]  4.67019122  3.14178647  7.92488795  0.00000000  0.00000000  2.86609781
#> [319]  4.75885618  7.80002107  1.28759566  3.12495397  4.33947244  2.15788392
#> [325]  3.64387304  6.77855761  5.02402903  0.89026760  1.39197057  0.00000000
#> [331]  4.53694097  4.76541644  4.48013500  0.60927467  4.93597393  5.84169475
#> [337]  2.94372350  2.38575271  2.37228075  0.35068072  3.90652966  0.37419725
#> [343]  7.61463038  4.13516906  2.24002913  1.95625477  2.05643258  3.43131165
#> [349]  2.79254787  3.37125233  1.89958344  4.22277097  4.32736400  6.46201365
#> [355]  1.67847494  2.77053430  2.94884331  5.19364734  1.56301008  0.00000000
#> [361]  0.70156762  9.09367925  4.42300500  0.00000000  0.00000000  7.14284301
#> [367]  3.92444571  3.68931155  7.20638380  0.21036136  1.46364888  1.82039829
#> [373]  4.14642907  0.00000000  1.93877412  4.25679172  3.95946927  4.97637327
#> [379]  5.10146715  2.87860291  3.60535970  0.13707193  2.50410238  0.00000000
#> [385]  5.11595398  7.54652385  5.30333451  3.46184124  5.64841699  4.71690326
#> [391]  2.43544312  1.00623176  4.17506871  1.43476439  1.55363716  4.91519404
#> [397]  4.02381874  2.32017539  3.86972173  0.00000000  7.12935062  3.92111732
#> [403]  0.87251638  6.38709421  7.07080308  4.07593345  0.00000000  0.97976954
#> [409]  8.37018197  3.98012436  8.30045381  3.84282412  0.00000000  4.81561620
#> [415]  0.61547818  7.14904666  1.15786408  1.45312037  0.00000000  4.04082918
#> [421]  5.84451991  0.62830598  3.77422116  2.25277407  1.51572934  5.55191491
#> [427]  6.17715302  1.31927730  6.54723208  2.31321971  9.97853320 10.14674473
#> [433]  2.73284879  4.40687510  1.50258485  1.41181638  5.55002580  1.73217539
#> [439]  3.27771935  0.16318692  1.72681325  6.07083383  5.02668017  4.69131169
#> [445]  4.66027175  4.61040855  6.61016929 10.57629517  0.75919377  7.32679827
#> [451]  0.00000000  3.61442662  1.98928342  0.00000000  5.75735489  0.84979367
#> [457]  3.90955634  4.77884565  0.72458075  0.59271279  2.82025166  0.56532811
#> [463]  1.89993762  6.09917880  2.01985718  0.00000000  7.35368945  7.13534052
#> [469]  3.55360123  8.60074333  7.53234606  4.24174851  3.21244702  3.18231569
#> [475]  0.00000000  0.00000000  8.87429244  1.09689626  1.82620046  5.76835247
#> [481]  3.30303716  7.71147530  5.91836933  3.12399669  1.80673237  6.32233398
#> [487]  7.30942684  7.29511547  5.59859121  2.06320314  0.00000000  0.78906254
#> [493]  1.45196669  4.62379897  5.92935326  0.41806432  4.31361415  4.61649016
#> [499]  0.00000000  0.00000000