Jones Model

Quantifies the chill accumulation by means of converting temperatures to exponential units temperature.

jones_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill for each temperature (TRUE is default).

Value

The function returns values the chill for each temperature of vector (Total = FALSE), or returns the chill accumulation (Total = TRUE).

References

JONES, H. G., HILLIS, R. M., GORDON, S. L., and BRENNAN, R. M. (2013). An approach to the determination of winter chill requirements for different Ribes cultivars. Plant. Biol.

Examples


 x <- rnorm(500, 8, 5)
jones_model(x)

#> [1] 142.5966

jones_model(x, FALSE)

#>   [1] 0.05026761 0.37265969 0.57643656 0.42198061 0.14866093 0.35742618
#>   [7] 0.17949801 0.13101796 0.36265490 0.02892350 0.12113493 0.23102205
#>  [13] 0.06193357 0.14631167 0.05789659 1.20271282 0.39939775 0.30717914
#>  [19] 0.14627848 0.12100458 0.31280499 0.12645740 0.29955789 0.22625364
#>  [25] 0.17308338 0.28133835 0.47993635 0.32598981 0.11343344 0.42161329
#>  [31] 0.34692014 0.20238849 0.59760131 0.10240302 0.20049562 0.10787504
#>  [37] 0.31666675 0.24179963 0.22737220 0.12431860 0.09111442 0.20593765
#>  [43] 0.08249008 0.22432144 0.30599451 0.10056680 0.35606202 1.24511530
#>  [49] 0.08018442 0.25792668 0.26853928 0.42287141 0.12611195 0.36855055
#>  [55] 0.05433594 0.04238270 0.32666352 0.13675318 0.45669433 0.19622496
#>  [61] 0.14124818 0.36730357 0.66457144 0.08938681 0.11016179 0.19409319
#>  [67] 1.62976886 0.72268490 0.10489686 0.15416772 0.51196205 0.77355185
#>  [73] 0.18242847 0.30789440 0.21766133 0.32508487 0.34972178 0.14897563
#>  [79] 0.09818308 0.30271825 0.29245758 0.40358261 0.20157818 0.10451001
#>  [85] 0.18811587 0.27668175 0.11963584 0.29024497 0.71893729 0.16609105
#>  [91] 0.31868414 0.42835981 0.23296681 0.13802041 0.28832521 0.37738455
#>  [97] 0.18980318 0.20493127 0.28650615 0.11341288 0.24469261 0.20002102
#> [103] 0.24895041 0.24656814 1.21150495 0.25891254 0.26555174 0.36952661
#> [109] 0.12680181 0.17215535 0.05376111 0.19750751 0.12453196 0.57528372
#> [115] 0.08656692 0.66620791 0.19923169 0.16582627 0.33458800 0.40815964
#> [121] 0.44322560 0.10148293 0.10311668 0.09025140 0.14300654 0.16137930
#> [127] 0.26296143 0.15336403 0.11908783 0.10288026 0.40865937 0.76686404
#> [133] 0.34053346 0.10203819 0.37290855 0.28594145 0.07140074 0.28643567
#> [139] 0.23347260 0.07303535 0.11743925 0.13559284 0.04400691 0.16520462
#> [145] 0.94034973 0.12066008 0.32732996 0.08043017 0.18378420 0.71280072
#> [151] 0.08512910 0.60148306 0.03889741 0.12282250 0.08701875 0.22919278
#> [157] 0.11555721 0.90374905 0.46388965 0.14519935 0.16483600 0.91409800
#> [163] 0.09249899 0.13968036 0.14523363 0.64686276 0.21558821 0.51540907
#> [169] 0.12489441 0.06862285 0.26591437 0.22253393 0.43553292 0.21613121
#> [175] 0.20620726 0.16387394 0.10542935 0.37497459 0.05047734 0.15091538
#> [181] 0.75872726 0.13268139 0.10538919 0.34432588 0.36829283 0.12062476
#> [187] 0.32967150 0.26536874 0.21911051 0.14872219 0.17850260 0.40214005
#> [193] 0.39337101 0.60043021 0.14659103 0.44239604 0.13221361 0.12458696
#> [199] 0.13217591 0.09672689 0.19876987 0.19447885 0.16556207 0.16047690
#> [205] 0.11915530 0.24775711 0.45377145 0.19782644 0.24734741 0.62237391
#> [211] 0.10841485 0.65348917 0.11453790 0.05540280 0.16757306 0.07869222
#> [217] 0.17142495 0.22371428 0.11845196 0.09289068 0.16327885 0.12100938
#> [223] 0.09646415 1.13905155 0.10551075 0.25420299 0.35749603 0.11854241
#> [229] 0.19142059 0.79159066 0.45328795 0.11103754 0.18772284 0.16732542
#> [235] 0.25659450 0.32524339 0.25811243 0.35533874 1.22776350 0.23004078
#> [241] 0.12165174 0.05967464 0.35479666 0.19977651 0.20902483 0.07253219
#> [247] 0.06354399 0.64151136 0.58336833 2.11626727 0.38429775 0.05115124
#> [253] 0.06890437 0.25937760 0.13986761 1.03235126 0.33873745 0.49626860
#> [259] 0.13448698 0.56779776 0.19255090 0.20438386 0.10736496 0.03164161
#> [265] 0.12472716 0.27126972 0.11177200 0.51085985 0.10908441 0.57899647
#> [271] 0.29751240 1.19653695 0.09599138 0.08357314 0.30211507 0.24069062
#> [277] 0.19822693 0.77462457 0.21887290 0.23149603 0.09997883 0.09709385
#> [283] 0.12351363 0.31057691 0.13422889 0.82484410 0.22779997 0.25800442
#> [289] 0.11194595 0.05691050 0.06230506 0.28655066 0.06269736 0.40032985
#> [295] 0.30180681 0.05310741 0.42962721 0.22906677 0.28999038 0.40459455
#> [301] 0.08323011 0.21872736 0.13843250 0.08684964 0.42293957 0.40970258
#> [307] 0.13956393 0.08442725 0.12524712 0.08808731 0.12082838 0.14700100
#> [313] 0.13524818 0.18613800 0.26277007 1.03117892 1.10263125 0.06259601
#> [319] 0.38811490 0.48047749 0.11548724 0.92694276 0.31875814 0.25521058
#> [325] 0.43313255 0.45120833 1.81784813 1.63414130 0.22305117 0.24563203
#> [331] 0.23863975 0.14562714 0.11335721 0.08853340 0.08503550 0.13422571
#> [337] 0.92599994 0.31891470 0.15046406 0.24732412 0.17193995 0.40804336
#> [343] 0.14857828 0.23255834 0.21227336 0.27948866 0.15229103 0.32552537
#> [349] 0.23251420 0.38058662 0.16518133 0.15973039 0.21670885 0.05549891
#> [355] 0.16227661 0.08653002 0.31232458 0.66197159 0.11193589 0.29443617
#> [361] 0.25600020 0.40695846 0.09565718 0.11556386 0.34255868 0.18855207
#> [367] 0.28517479 0.07608419 0.51271589 0.31081759 0.11368882 0.70989989
#> [373] 0.18133146 0.12457647 0.18900758 0.14761924 0.24541699 0.22792687
#> [379] 0.32111359 0.34458850 0.41396807 0.09747653 0.40008029 0.20271311
#> [385] 0.22981130 0.48689032 0.10103062 0.27113751 0.17029515 0.42735162
#> [391] 0.18668732 0.14286182 0.29688130 0.07427574 0.31212991 0.54680763
#> [397] 0.24121287 0.18949786 0.43823104 0.47628680 0.40354918 0.52928903
#> [403] 0.39732965 0.25478087 0.21307801 0.20107339 0.31726607 0.10494515
#> [409] 0.22228609 0.06333179 0.27363401 0.08697170 0.05745158 0.25954704
#> [415] 0.10326384 0.32053177 0.42290305 0.13689496 0.25494423 0.61992999
#> [421] 0.18176560 0.33570065 0.22358856 0.10517480 0.51838733 0.36781152
#> [427] 0.10897001 0.10792432 0.46664032 0.19497223 1.22626471 0.13457540
#> [433] 0.09475054 0.27508130 0.16351877 0.11922771 0.30307129 0.45032309
#> [439] 0.17747650 0.31370826 0.16514169 0.65005263 0.17885687 0.37544905
#> [445] 0.27321971 0.30433213 0.45289332 0.22078502 0.19956610 0.03288146
#> [451] 0.40678754 0.25490493 0.21378366 0.27095074 0.45104719 0.07302195
#> [457] 0.67183820 0.12266745 0.09307906 0.18018057 0.57777612 0.17261436
#> [463] 0.35632550 0.14339083 0.37422068 0.15402925 0.15197501 0.35596972
#> [469] 0.23262092 0.09152576 0.07884187 1.06422417 0.24931276 0.16552941
#> [475] 0.21896156 0.15447765 0.05891666 0.56911848 0.73123981 0.36935638
#> [481] 0.28417428 0.15748295 0.04167693 0.23837278 0.12753809 0.13210694
#> [487] 0.12594445 0.21524602 0.22640270 0.30276930 0.14705804 0.08992988
#> [493] 0.29213906 0.16405548 0.10890844 0.51064050 0.25077920 0.05482509
#> [499] 0.13773820 0.11532178