Landsberg Model

Quantifies the chill accumulation based in the base temperature.

landsberg_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).

Value

The function returns values the chill for each temperature of vector (Total = FALSE), or returns the chill accumulation (Total = TRUE).

Details

The model is based on the subtraction of the base temperature of each hourly temperature. The temperature of the base is 5°C.

References

LANDSBERG J. J. (1974). Apple Fruit Bud Development and Growth; Analysis and an Empirical Model. Annals of Botany.

Examples


x <- rnorm(500, 8, 5)
landsberg_model(x)

#> [1] 324.0998

landsberg_model(x, FALSE)

#>   [1] 1.0000000 0.4214070 0.4790346 0.5623053 1.0000000 0.7427199 0.4468068
#>   [8] 0.5044594 1.0000000 1.0000000 1.0000000 0.3622745 0.4865787 0.3912629
#>  [15] 0.5103696 0.6364656 1.0000000 1.0000000 0.6285437 0.4775132 0.5410025
#>  [22] 0.5715516 0.2734457 0.8321671 1.0000000 0.6447999 0.4199416 0.2463895
#>  [29] 0.3843858 1.0000000 1.0000000 1.0000000 0.8737140 1.0000000 1.0000000
#>  [36] 1.0000000 1.0000000 0.5279450 0.2514539 1.0000000 0.9227380 1.0000000
#>  [43] 0.7016343 1.0000000 0.4361624 0.8326312 1.0000000 1.0000000 1.0000000
#>  [50] 0.5122978 1.0000000 1.0000000 0.3815048 0.6244983 0.3872596 1.0000000
#>  [57] 0.6738196 0.2992517 1.0000000 0.4707138 0.4491610 1.0000000 0.5987812
#>  [64] 0.5019061 0.5820942 0.3412117 0.5472329 0.9145911 0.4884874 0.3098750
#>  [71] 0.7576009 0.3653766 0.4108367 1.0000000 1.0000000 0.6450152 0.3007642
#>  [78] 0.7000706 0.4682227 0.4228557 1.0000000 0.5240872 0.5031697 1.0000000
#>  [85] 0.3725799 1.0000000 0.4018360 1.0000000 0.4623755 0.8196215 0.8756029
#>  [92] 0.6897305 1.0000000 0.7277705 0.5539721 0.6958905 0.6626598 0.2743635
#>  [99] 0.6415259 1.0000000 0.4858845 0.5634408 0.4626164 0.3610863 0.6887421
#> [106] 0.5178403 0.4934622 0.5409430 1.0000000 0.4862795 0.4863397 0.2986133
#> [113] 0.6069445 0.4757184 0.6227744 1.0000000 0.8062082 0.4580796 0.5329116
#> [120] 0.4566257 0.3849344 0.3446262 0.4213445 0.3795523 0.3808839 1.0000000
#> [127] 0.8253460 0.7958053 0.5741767 0.4695105 0.5653232 0.6798433 0.7163045
#> [134] 0.3921711 1.0000000 0.3656420 0.5191201 0.8847392 0.4276159 0.3690579
#> [141] 0.2962553 0.6699990 0.3466355 0.3730823 0.5291597 0.8172020 0.2654587
#> [148] 0.4491766 0.7471457 0.3838492 0.9128570 1.0000000 0.5123116 0.6832625
#> [155] 0.7255071 1.0000000 0.3454222 0.6531022 0.4943375 1.0000000 0.3996206
#> [162] 0.6201593 1.0000000 0.3344996 0.6795578 0.8928346 1.0000000 0.7148678
#> [169] 0.4328008 0.7341009 0.5549607 0.3538869 0.8412886 0.7011618 0.4920437
#> [176] 1.0000000 0.3234886 0.9282994 0.3454784 1.0000000 0.5285770 0.3328657
#> [183] 0.4986258 0.7566827 1.0000000 0.3284719 0.5220022 0.4635003 0.3904030
#> [190] 0.5159961 0.8837775 0.5641633 0.6062686 0.4496930 0.3340989 1.0000000
#> [197] 0.4219263 0.4016279 0.4545712 1.0000000 1.0000000 0.5817034 0.7882641
#> [204] 0.4341064 0.4698383 0.7277664 1.0000000 1.0000000 0.8784638 0.4502559
#> [211] 0.3154279 1.0000000 0.5213581 0.6150532 1.0000000 1.0000000 0.3804566
#> [218] 0.4734910 0.3177061 0.2938309 1.0000000 0.5462492 0.3519722 0.5653928
#> [225] 0.3604734 0.5688281 0.4209545 0.6195436 0.7590885 1.0000000 0.3325118
#> [232] 0.5471834 0.6938917 0.4864154 0.4349280 0.3269379 0.3050163 0.8625574
#> [239] 0.5137700 1.0000000 0.5353246 0.5033479 0.4502318 0.6298258 0.7276234
#> [246] 0.2744975 1.0000000 0.6190428 0.4772839 1.0000000 1.0000000 1.0000000
#> [253] 0.4445199 0.5412276 1.0000000 0.4453404 1.0000000 1.0000000 1.0000000
#> [260] 1.0000000 1.0000000 1.0000000 0.3987404 0.4863805 0.4000401 0.3194877
#> [267] 0.6641161 1.0000000 0.4040314 0.3454109 0.6746014 0.4685265 0.4995846
#> [274] 0.8022786 0.8440835 0.2827863 0.6626512 0.7672446 0.6606024 0.5346912
#> [281] 0.3805459 0.5382246 0.5085381 0.3258582 1.0000000 0.9030756 0.5043676
#> [288] 0.4796996 0.9295914 0.6079089 0.5568934 1.0000000 0.4091002 1.0000000
#> [295] 0.2921463 0.6410838 1.0000000 0.4446642 1.0000000 0.8054841 0.4151313
#> [302] 0.2944577 0.2133610 0.5186812 0.6691921 0.4746914 0.3610144 0.4247388
#> [309] 1.0000000 0.6704275 0.4208525 0.4533926 0.4801282 0.3048306 0.6900450
#> [316] 0.4386735 0.4878286 0.5954394 0.5138084 0.5127459 0.6087415 0.5821038
#> [323] 0.8272846 0.4082055 0.4461761 1.0000000 0.3278423 0.9837462 1.0000000
#> [330] 0.3277913 1.0000000 0.4818888 1.0000000 0.5115572 0.4815534 0.3703405
#> [337] 0.6694813 1.0000000 1.0000000 0.5236125 1.0000000 0.3740120 0.4723653
#> [344] 1.0000000 1.0000000 0.4921997 0.3782209 0.3470459 0.7809236 0.4980089
#> [351] 1.0000000 0.7583933 0.6882343 0.5587514 0.4439541 0.6669871 0.4065839
#> [358] 1.0000000 0.4675593 0.2939581 0.7310862 0.4241009 0.4646238 1.0000000
#> [365] 0.7984867 1.0000000 1.0000000 0.6431525 0.7171761 1.0000000 1.0000000
#> [372] 0.3621647 1.0000000 0.9978795 1.0000000 0.9090782 0.2543237 0.7500196
#> [379] 1.0000000 0.4442460 0.3573503 0.5053752 0.6524756 0.4628771 0.2780880
#> [386] 0.6562063 0.4152729 0.9227812 0.6055262 0.5528916 1.0000000 0.2930066
#> [393] 1.0000000 0.3220174 0.7538088 0.9101635 0.8179568 0.5119714 0.4622878
#> [400] 0.3042698 0.9677022 0.8368285 0.9268406 0.3321656 0.4606394 0.3767188
#> [407] 0.4411143 0.4997611 1.0000000 0.4330601 0.6350634 0.4168613 1.0000000
#> [414] 0.4355985 1.0000000 0.4193469 0.4051111 0.4434428 0.9054056 1.0000000
#> [421] 0.5014208 0.6207331 0.3040776 0.4840801 0.3241090 0.8719906 0.3938180
#> [428] 0.5078530 1.0000000 0.9376358 1.0000000 0.4169273 1.0000000 0.5383399
#> [435] 0.9896309 0.4481404 1.0000000 0.9839608 0.6657354 1.0000000 0.3176124
#> [442] 0.2428127 0.3114539 0.5886876 0.8326712 0.8278097 0.3692013 0.7283860
#> [449] 1.0000000 0.5763757 1.0000000 0.5388209 0.5023999 1.0000000 0.4717971
#> [456] 1.0000000 1.0000000 1.0000000 0.4204445 1.0000000 0.5625940 0.4652987
#> [463] 0.9324023 1.0000000 0.4685204 1.0000000 0.6364166 1.0000000 0.3670548
#> [470] 0.8090345 0.4276985 0.5216691 0.6180546 0.2221194 0.3704385 0.7482403
#> [477] 0.4127489 0.3781881 0.4454715 1.0000000 0.5039149 0.5703258 0.4265946
#> [484] 0.5166712 0.3002994 0.3646728 0.6461958 0.8382147 1.0000000 1.0000000
#> [491] 1.0000000 0.4581415 0.3621874 0.8518365 0.6048579 0.3953500 1.0000000
#> [498] 1.0000000 0.7725949 0.6017677