North Carolina

Quantifies the chill accumulation by means of converting temperatures to chill-units.

north_carolina(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).

Value

The function returns values the chill-units for each temperature of vector (Total = FALSE), or returns the chill-units accumulation (Total = TRUE).

Details

The model is based on chill-units and optimum temperature, minimum and maximum temperature limits (They aren't classes), where 1 chill-unit is when the tree is exposure at optimum temperature (7.2°C). When the temperature is above 23°C (maximum temperature limit), the chill-unit is -2. The chill-units accumulation is 0 when occurs temperature below -1.1°C, being the miminum temperature limit.

References

SHALTOUT, Assem D. & UNRATH, C. R. 1983. Rest Completion Prediction Model for 'Starkrimson Delicious' Apples. J. Amer. Soc. Hort. Sci.

Examples


x <- rnorm(500, 5, 4)
north_carolina(x)

#> [1] 349.491

north_carolina(x, FALSE)

#>   [1]  0.70917901  0.92503797  0.89419808  0.49225081  0.92670650  0.48814336
#>   [7]  0.94014892  0.64610911  0.58055066  0.86645329  0.13854721  0.82807132
#>  [13]  0.70387966  0.95010315  0.63944136  0.92504417  0.93244151  0.95872588
#>  [19]  0.81557781  0.22977884  0.79076865  0.91486670  0.99445635  0.65245787
#>  [25]  0.30986674  0.92482063  0.96150062  0.00000000  0.00000000  0.99504232
#>  [31]  0.96379368  0.95322913  0.97974504  0.88567036  0.80963883  0.18125044
#>  [37]  0.55800614  0.67084498  0.59316804  0.35608756  0.82423550  0.96461850
#>  [43]  0.94291144  0.96288601  0.97834654  0.49691522  0.52007298  0.27893183
#>  [49]  0.62260674  0.93952520  0.32761141  0.76712388  0.08535464  0.92209052
#>  [55]  0.95913056  0.69447655  0.97574391  0.45824598  0.55483893  0.91389643
#>  [61]  0.33853463  0.72561831  0.70576425  0.36570682  0.05151398  0.80750411
#>  [67]  0.89851122  0.88712018  0.69034008  0.98912123  0.78326084  0.00000000
#>  [73]  0.87880271  0.61516787  0.87502221  0.90121738  0.85603724  0.88133745
#>  [79]  0.92796792  0.94673630  0.95761987  0.93340578  0.89206695  0.64326937
#>  [85]  0.96316439  0.85470712  0.95945132  0.39297439  0.95571947  0.29843083
#>  [91]  0.91470430  0.59643166  0.83590225  0.85376729  0.61003916  0.76927427
#>  [97]  0.77100125  0.43704882  0.90691076  0.36005913  0.96207232  0.91887410
#> [103]  0.60605395  0.97340186  0.53573083  0.00000000  0.04190687  0.94059556
#> [109]  0.84781332  0.79502493  0.81054388  0.62479655  0.73324990  0.54566871
#> [115]  0.75262018  0.00000000  0.93940272  0.97226914  0.95675721  0.94677118
#> [121]  0.37273761  0.95552795  0.89714735  0.90118084  0.95310691  0.51524461
#> [127]  0.95976649  0.93755202  0.67733006  0.18720195  0.88527888  0.81227311
#> [133]  0.25591159  0.44174776  0.65075023  0.90460927  0.00000000  0.90418558
#> [139]  0.71329317  0.41172996  0.89529593  0.97194269  0.29048070  0.70422649
#> [145]  0.92113716  0.75395108  0.93033851  0.52276639  0.95496453  0.71790561
#> [151]  0.80434748  0.93706793  0.82233630  0.73904301  0.87565627  0.94253048
#> [157]  0.95009366  0.77022884  0.15370745  0.91735028  0.00000000  0.88098057
#> [163]  0.96232525  0.95572011  0.98226629  0.77650717  0.64688319  0.43667952
#> [169]  0.77613807  0.83961609  0.04333057  0.66514905  0.66867786  0.83509268
#> [175]  0.93050478  0.98347769  0.00000000  0.00000000  0.90445226  0.55449895
#> [181]  0.94002749  0.92355820  0.94605085  0.95749502  0.75270504  0.94390601
#> [187]  0.74361659  0.81893309  0.90844001  0.97353066  0.99389382  0.95996364
#> [193]  0.95540762  0.94754166  0.71114197  0.86439311  0.86864488  0.88498822
#> [199]  0.69924339  0.65417946  0.50080997  0.90694094  0.73000653  0.78260455
#> [205]  0.59941911  0.00000000  0.82074498  0.72410841  0.96408568  0.45277956
#> [211]  0.88005393  0.85208332  0.92198407  0.22132950  0.95406616  0.94559188
#> [217]  0.66624484  0.94815261  0.71740966  0.05480895  0.94435076  0.92981267
#> [223]  0.98148719  0.65813389  0.93987311  0.85413779  0.40628441  0.59998845
#> [229]  0.42091197  0.98978195  0.92283523  0.00000000  0.98111126  0.93301079
#> [235]  0.76797772 -0.10432124  0.11051671  0.04873574  0.87600296  0.48517795
#> [241]  0.96420214  0.74077026  0.17868681  0.92375653  0.74426304  0.91708529
#> [247]  0.73604811  0.74111400  0.71300636  0.71011729  0.73137086  0.00000000
#> [253]  0.69508876  0.76751659  0.54351795  0.12179976  0.96162602  0.97577034
#> [259]  0.72488119  0.95665575  0.95882159  0.92018456  0.70070058  0.00000000
#> [265]  0.93628654  0.54380628  0.59822333  0.95436101  0.54634116  0.00000000
#> [271]  0.49301173  0.87994806  0.96378624  0.91558242  0.91649181  0.00000000
#> [277]  0.81417283  0.87676008  0.95894827  0.96125909  0.96531212  0.51187450
#> [283]  0.78953599  0.79614231  0.84587298  0.85575480  0.86970674  0.84160397
#> [289]  0.64904939  0.00000000  0.91904414  0.96706600  0.81866508  0.68342469
#> [295]  0.72592614  0.94922760  0.88552935  0.83033783  0.90683867  0.77360864
#> [301]  0.89298906  0.68984545  0.93959531  0.86000261  0.95795338  0.62256398
#> [307]  0.68266079  0.93808144  0.11698686  0.94387699  0.98380115  0.89818903
#> [313]  0.00000000  0.00000000  0.54674472  0.94382898  0.89585571  0.00000000
#> [319]  0.96134056  0.58840884  0.95872586  0.46565865  0.00000000  0.24272994
#> [325]  0.91514797  0.05270571  0.85301421  0.96319765  0.96292019  0.99283270
#> [331]  0.92674108  0.91441304  0.85950367  0.04449793  0.87811747  0.97716116
#> [337]  0.88068609  0.68614338  0.86242839  0.33469371  0.66574917  0.48029910
#> [343]  0.87561135  0.00000000  0.39861368  0.98859771  0.71467020  0.96326936
#> [349]  0.70955290  0.95573132  0.88356943  0.78043435  0.94955243  0.82360854
#> [355]  0.31325971  0.00000000  0.90338695  0.97501905  0.89283494  0.60264834
#> [361]  0.64241323  0.94181732  0.30027268  0.98383574  0.56856805  0.96258866
#> [367]  0.96185620  0.91321804  0.35922231  0.94156039  0.00000000  0.92234510
#> [373]  0.52613317  0.95117055  0.96094918  0.00000000  0.96934303  0.84480714
#> [379]  0.91859815  0.10806835  0.71929675  0.59372945  0.40537606  0.81699302
#> [385]  0.41490351  0.45122713  0.45870066  0.24547108  0.28367705  0.95960916
#> [391]  0.68417185  0.96230303  0.00000000  0.31760670  0.95142502  0.95799028
#> [397]  0.95648155  0.57573535  0.99680860  0.45636364  0.48182416  0.52486789
#> [403]  0.95941195  0.76426834  0.12397731  0.29497054  0.00000000  0.64114703
#> [409]  0.78130157  0.56034514  0.00000000  0.88907181  0.85518785  0.31820727
#> [415]  0.08214757  0.55900774  0.92246174  0.90065253  0.91390251  0.94124770
#> [421]  0.95492474  0.21791216  0.58398802  0.86525370  0.00000000  0.98905340
#> [427]  0.74593864  0.00000000  0.83572482  0.26125731  0.95087396  0.96245767
#> [433]  0.48883973  0.33715960  0.98745737  0.92435672  0.95971980  0.75671756
#> [439]  0.41457661  0.63671465  0.79240958  0.93716496  0.96130322  0.95534859
#> [445]  0.98881515  0.96365511  0.46340096  0.54011769  0.99211654  0.42145602
#> [451]  0.98853362  0.84730464  0.57889874  0.64021653  0.94163114  0.56837088
#> [457]  0.97964626  0.64746879  0.58683257  0.84732365  0.88382955  0.97202794
#> [463]  0.83819984  0.84661577  0.83044009  0.68313176  0.34521205  0.86269972
#> [469]  0.16043367  0.99628481  0.93718529  0.00000000  0.94513994  0.71616618
#> [475]  0.95277420  0.81348982  0.49716133  0.91295724  0.99172502  0.97286480
#> [481]  0.41453613  0.90034942  0.90706975  0.23315401  0.71688292  0.46319181
#> [487]  0.19169640  0.94011614  0.95798435  0.93050925  0.96723984  0.70067499
#> [493]  0.79296115  0.97420835  0.53842636  0.27305811  0.91427904  0.81086289
#> [499]  0.41002308  0.00000000