Positive Chill Units - PCU

Quantifies the chill accumulation by means of converting temperatures to positive chill-units - Modified Utah Model.

pcu_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).

Value

The function returns values the chill-units for each temperature of vector.

Details

The PCU Model is the modified Utah Model. When the temperature is above 15.9°C, the chill-unit is 0. This modification was made because when the Utah model is applied in warm conditions, accumulation becomes negative.

References

Richardson, E. A. et al. 1974. "A Model for Estimating the Completation of Rest for 'Redhaven' and 'Elberta' Peach Trees". Research Reports & Notes.

Linsley-Noakes, G. C. et al. 1995. "Estimating daily positive Utah Chill units using daily minimum and maximum temperatures".

Examples


x <- rnorm(500, 15, 4)
pcu_model(x)

#> [1] 85.5

pcu_model(x, FALSE)

#>   [1] 0.5 0.0 0.0 0.5 0.0 0.0 0.5 0.0 0.5 0.0 0.5 0.0 0.0 0.5 0.0 0.0 0.0 0.5
#>  [19] 0.0 0.5 1.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.5 0.0 0.0
#>  [37] 0.0 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.0
#>  [55] 0.0 0.0 0.5 0.0 1.0 0.0 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.5
#>  [73] 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.5 0.0 0.5 0.0
#>  [91] 0.5 0.0 0.5 0.5 0.0 0.0 0.0 0.0 0.5 0.0 0.0 1.0 0.5 0.0 0.5 0.0 0.0 0.0
#> [109] 0.0 0.0 0.0 0.0 0.0 0.0 0.5 1.0 0.0 0.0 0.5 0.5 0.0 0.0 0.0 0.0 0.0 0.5
#> [127] 0.0 0.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
#> [145] 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.0 0.5 0.0
#> [163] 0.0 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.0 0.5 0.0 0.0 0.5 0.0 0.0 1.0 0.0 0.0
#> [181] 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.5 0.0 0.0
#> [199] 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.0 1.0 0.0 0.5 0.0 0.0 0.0 0.0
#> [217] 0.0 0.0 0.0 1.0 1.0 0.5 0.0 0.0 0.5 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0
#> [235] 0.0 0.0 0.5 0.5 0.0 0.0 0.0 0.0 0.5 0.0 0.5 1.0 0.0 1.0 0.0 0.0 0.0 1.0
#> [253] 0.5 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.5 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.5
#> [271] 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.5 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.0 0.0
#> [289] 0.0 0.5 0.0 0.0 0.0 1.0 0.0 0.0 0.5 0.0 0.5 0.5 0.0 0.0 0.5 0.0 0.0 0.0
#> [307] 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.5 0.0 0.5 0.5 0.0 0.0 1.0 0.5 0.0 0.5 0.0
#> [325] 0.0 0.0 0.0 0.0 0.5 0.5 1.0 0.5 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.0 0.5
#> [343] 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.5 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
#> [361] 0.5 0.0 0.0 0.5 0.5 0.0 0.0 0.0 0.0 0.0 0.5 0.5 0.0 0.5 0.0 0.5 0.0 0.0
#> [379] 0.0 0.5 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
#> [397] 0.5 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0
#> [415] 0.0 0.0 1.0 0.0 0.5 0.0 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
#> [433] 0.0 0.0 0.0 1.0 0.5 0.5 0.0 0.0 0.0 0.5 0.0 0.0 0.0 0.5 0.5 1.0 0.0 0.0
#> [451] 0.0 0.0 0.5 0.0 0.5 0.0 0.0 0.0 0.0 1.0 0.0 0.5 0.0 1.0 0.0 1.0 0.0 0.0
#> [469] 0.0 0.5 1.0 0.0 0.0 0.0 0.0 0.5 0.0 0.0 0.0 0.0 0.0 0.5 1.0 0.0 1.0 0.0
#> [487] 0.0 1.0 0.0 0.5 0.0 0.0 0.5 0.0 0.0 1.0 1.0 0.0 0.0 0.0