Bidabe Model

Quantifies the chill accumulation by means of converting temperatures to temperature coeficients (Q10).

q10_bidabe(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).

Value

The function returns values for each temperature of vector (Total = FALSE), or returns the accumulation (Total = TRUE).

Details

The model is based on temperature coeficients (Q10). The Q10 coeficient is variable for each specie. The model contains dinamic variables. This model is been aplied in the warm climates. Was based on an exponential function that decreases according to the increase in temperature.

References

BIDABE, B. 1967. Action de la temperature sur l' evolution des bourgeons de pommier et comparaison de methodes de controle de l' epoque de floraison. Annu. Physiol. Veg.

Examples


x <- rnorm(500, 10, 3)
q10_bidabe(x)

#> [1] 7.188525

q10_bidabe(x, FALSE)

#>   [1] 0.010759901 0.037311645 0.010689622 0.009812112 0.014216849 0.017213026
#>   [7] 0.012648609 0.016063360 0.007300084 0.018578918 0.015722997 0.013669914
#>  [13] 0.013035744 0.017938408 0.010546271 0.018042567 0.017687459 0.010718202
#>  [19] 0.013832713 0.015700967 0.031225121 0.009009691 0.021235643 0.020501581
#>  [25] 0.009732953 0.017402762 0.008359260 0.013827232 0.014813482 0.015151715
#>  [31] 0.009219255 0.015861286 0.006119191 0.019854858 0.016818092 0.015483716
#>  [37] 0.011571660 0.015709429 0.010158762 0.023512722 0.012240925 0.016071439
#>  [43] 0.020650379 0.011187155 0.012280042 0.011517705 0.016823416 0.007952602
#>  [49] 0.019892048 0.011843262 0.020504205 0.018009919 0.014175252 0.009221109
#>  [55] 0.019416106 0.014878017 0.011061737 0.011578135 0.019744868 0.013360030
#>  [61] 0.010404689 0.020124967 0.013173369 0.021473447 0.020944593 0.016354955
#>  [67] 0.018920790 0.014014122 0.020484188 0.012645475 0.016210680 0.010518969
#>  [73] 0.010530468 0.011647097 0.011656891 0.008883259 0.013081416 0.008803338
#>  [79] 0.012231775 0.012904993 0.011425069 0.016610439 0.024260072 0.012303676
#>  [85] 0.013729362 0.021827578 0.017107781 0.012455004 0.010290744 0.019084853
#>  [91] 0.014692743 0.014680365 0.012861582 0.018820253 0.008921070 0.014093029
#>  [97] 0.009612215 0.022989313 0.014134950 0.012455398 0.019185858 0.022678515
#> [103] 0.013105560 0.010471135 0.013523837 0.015852640 0.011792563 0.012016108
#> [109] 0.013244758 0.019169677 0.008988189 0.024521387 0.012528032 0.021085574
#> [115] 0.012197969 0.020329927 0.015103305 0.018404498 0.013429855 0.012811472
#> [121] 0.012912629 0.013064982 0.009641505 0.015282252 0.011280955 0.011413575
#> [127] 0.015131712 0.007118375 0.012088396 0.011236938 0.014936454 0.015397146
#> [133] 0.019988537 0.014192970 0.013779470 0.013787984 0.017045212 0.016113453
#> [139] 0.012165857 0.023723360 0.018584195 0.006235244 0.027645081 0.007723650
#> [145] 0.011940397 0.008308410 0.026282562 0.013353494 0.012562836 0.007230049
#> [151] 0.017417707 0.016370647 0.017704246 0.007618036 0.010972007 0.020974617
#> [157] 0.005540940 0.018063222 0.015331998 0.008434320 0.015036182 0.013244510
#> [163] 0.010027848 0.013994457 0.015451816 0.012199501 0.017593347 0.020168362
#> [169] 0.008325977 0.014490469 0.019521495 0.011031691 0.016427485 0.014812588
#> [175] 0.012343169 0.011676664 0.030153757 0.010081965 0.012875685 0.014912429
#> [181] 0.012156008 0.011949273 0.024451311 0.010772007 0.034352599 0.018743264
#> [187] 0.018508038 0.023648750 0.007465641 0.021422344 0.011869475 0.015325072
#> [193] 0.006093508 0.015412781 0.026802126 0.015637379 0.010137396 0.010833800
#> [199] 0.014021818 0.017341086 0.012524897 0.013441277 0.011505936 0.016340945
#> [205] 0.010882874 0.020328845 0.011176522 0.015054290 0.019798598 0.012843884
#> [211] 0.016577560 0.012578657 0.017392940 0.019936300 0.013402957 0.011172988
#> [217] 0.014638608 0.010240671 0.010503652 0.012829750 0.011167947 0.017339892
#> [223] 0.013158888 0.017657166 0.010797826 0.021808388 0.011140288 0.012831579
#> [229] 0.012906369 0.023525154 0.013552856 0.015249692 0.011684255 0.011442811
#> [235] 0.015475839 0.013876300 0.016890900 0.019439724 0.012374777 0.009873381
#> [241] 0.008518492 0.010779078 0.013647850 0.019407797 0.007448465 0.009422509
#> [247] 0.011041173 0.014661295 0.013871089 0.009154481 0.011639448 0.008145318
#> [253] 0.008031328 0.014550380 0.014470369 0.008950382 0.009385350 0.007510308
#> [259] 0.022116002 0.013129767 0.015563770 0.013039336 0.014396858 0.028216147
#> [265] 0.011658569 0.007929724 0.014357540 0.010015163 0.013973067 0.011979146
#> [271] 0.019002778 0.015075007 0.012422088 0.016303040 0.018554503 0.010060225
#> [277] 0.011129648 0.012849524 0.006754848 0.020553070 0.009731825 0.010090759
#> [283] 0.007844695 0.016237888 0.010171635 0.010512389 0.015803073 0.009020264
#> [289] 0.013805751 0.009031115 0.010925692 0.015580930 0.015018317 0.011822473
#> [295] 0.015355981 0.029505727 0.012908431 0.006282554 0.013595344 0.011526759
#> [301] 0.010092947 0.018197420 0.015813373 0.028340641 0.010086310 0.013318190
#> [307] 0.018060989 0.016865353 0.008627656 0.010700056 0.011370802 0.013894092
#> [313] 0.008826410 0.012683769 0.017443162 0.010651519 0.012993730 0.013333593
#> [319] 0.009996218 0.015322071 0.019789363 0.009660150 0.015814149 0.011568053
#> [325] 0.013350354 0.011715766 0.009900575 0.010912893 0.014413294 0.008671632
#> [331] 0.010339925 0.022773709 0.017535319 0.014428071 0.013493812 0.012196703
#> [337] 0.016975193 0.017213473 0.008516917 0.012722194 0.016925084 0.008560401
#> [343] 0.017036826 0.011522088 0.008797547 0.008759801 0.024816593 0.016338177
#> [349] 0.010480273 0.012230030 0.005859844 0.014431508 0.010405136 0.009695036
#> [355] 0.007445441 0.012128263 0.018806601 0.018181773 0.011860879 0.016173351
#> [361] 0.010735346 0.014629037 0.018587944 0.009019509 0.017734069 0.020443226
#> [367] 0.015160710 0.009607143 0.024780494 0.022418087 0.020424953 0.019200052
#> [373] 0.015150714 0.010353062 0.013081567 0.018255280 0.012135382 0.014757706
#> [379] 0.015128902 0.023084201 0.010041124 0.014584451 0.010910551 0.016624747
#> [385] 0.015591410 0.008696244 0.022560075 0.016301250 0.020215895 0.013992653
#> [391] 0.017505087 0.014936779 0.019356366 0.011960890 0.012909645 0.013032253
#> [397] 0.017271040 0.013153546 0.015897749 0.012429939 0.009215394 0.009660801
#> [403] 0.010986770 0.024809406 0.013815508 0.010185455 0.015368642 0.014905056
#> [409] 0.014285433 0.014704312 0.014006089 0.014114022 0.015840944 0.008940523
#> [415] 0.015595460 0.014448704 0.009832329 0.009042474 0.011522802 0.012380904
#> [421] 0.012928998 0.019913614 0.015707930 0.013076967 0.014241309 0.016559968
#> [427] 0.020129427 0.025912530 0.016164733 0.019313761 0.026048733 0.011662164
#> [433] 0.014102858 0.007491505 0.009997037 0.014354705 0.012466700 0.011622470
#> [439] 0.011282840 0.011927809 0.018669850 0.011638592 0.014542896 0.015723275
#> [445] 0.014638320 0.018139893 0.011274848 0.015089106 0.010597990 0.012061152
#> [451] 0.019043460 0.026146542 0.016015820 0.019689635 0.013135148 0.024998590
#> [457] 0.013726400 0.019180499 0.009186400 0.020166369 0.006933676 0.018631187
#> [463] 0.009106807 0.015998385 0.019483664 0.010724724 0.013320840 0.013597897
#> [469] 0.011238605 0.013474397 0.009767244 0.012678813 0.012527813 0.009667060
#> [475] 0.009683801 0.006755372 0.020955455 0.013665630 0.011761123 0.007989553
#> [481] 0.014860822 0.016037432 0.013308368 0.012919813 0.009148073 0.011783802
#> [487] 0.015086785 0.016732360 0.017711919 0.020863552 0.007827958 0.016162110
#> [493] 0.009530448 0.010439366 0.006400586 0.021101107 0.019194478 0.017646253
#> [499] 0.019363280 0.013455612