Taiwan model

Quantifies the chill accumulation by means of converting temperatures to chill-units.

taiwan_model(x, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).

Value

The function returns values the chill-units for each temperature of vector (Total = FALSE), or returns the chill-units accumulation (Total = TRUE).

Details

The model is based on chill-units, where 1 chill-unit is when the tree is exposure below 7.2°C. When the temperature is between 15.1°C and 26.6°C, there isn't accumulation of chill-unit. The chill-units accumulation is negative when occurs temperature above 26.7°C, and the chill-unit is -1 when occurs temperature above 27.8°C.

References

LU, M. T. et al. 2012. A model for estimating chilling requirement of very low-chill peaches in Taiwan. Acta Horticulturae, n. 962, p. 245.

Examples


x <- rnorm(500, 20, 7)
taiwan_model(x)

#> [1] -11.5

taiwan_model(x, FALSE)

#>   [1]  0.0 -1.0  0.5  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0 -0.5  0.0  0.0
#>  [16]  0.5 -1.0  0.0  0.0 -1.0 -1.0 -1.0  0.0  0.0  0.0  0.5  0.0  0.0 -1.0  1.0
#>  [31]  0.0  0.0  0.0 -1.0  0.0  0.0  0.0  0.5  0.0 -1.0  0.0  0.5  0.0  0.5 -1.0
#>  [46]  0.0  0.0  0.0 -1.0  0.5  1.0  0.0 -0.5 -0.5 -0.5  0.0  0.0  1.0  0.0  0.0
#>  [61]  0.0  0.5  0.0  0.0  0.0  0.0 -1.0  1.0  0.0  0.0  0.0 -1.0  0.5  1.0  0.0
#>  [76]  0.0  0.0  1.0  0.5  0.0  0.5  0.5 -1.0  0.5 -1.0 -1.0  0.0  0.0  0.0  0.0
#>  [91]  0.5  0.0  0.5  0.0  0.5  0.0  0.5  0.5 -0.5 -1.0  0.0  0.5  0.0  0.0  0.5
#> [106]  0.5  0.0  0.0  0.0  0.0  0.5  0.0 -1.0  0.0  0.5  0.0  0.0  0.5  0.0  0.0
#> [121]  0.0  0.0  0.5  0.5  0.0  0.0 -1.0  0.0  0.0  0.0  0.0 -1.0  0.0  0.0  0.0
#> [136]  0.0  0.0  0.0 -1.0  0.0  0.0  0.5  0.5 -1.0  0.5 -0.5 -1.0  0.0  0.0  0.0
#> [151]  0.5  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.5  0.0 -1.0  0.0  0.5  0.0
#> [166]  0.5  0.0  0.0 -1.0  0.0  1.0 -0.5  0.0  0.0  0.0  0.0 -1.0  0.0  0.5  0.5
#> [181]  0.0  0.0 -1.0  0.0  0.0  0.5 -1.0  0.0  0.0  0.0  0.5  0.0  0.0  0.5  0.0
#> [196]  0.0 -1.0  1.0 -0.5 -1.0 -1.0  1.0  0.5  0.5  0.5  0.0  0.0  0.0  0.0  0.0
#> [211]  0.5  0.0  0.5  0.5  0.5  0.0 -1.0 -1.0  0.0  0.0  0.5  0.5  0.0  0.5  1.0
#> [226]  0.0  0.0  0.5  0.0 -1.0 -1.0 -1.0  0.5 -0.5 -0.5  0.0  0.0  0.5  0.5 -1.0
#> [241] -1.0 -1.0  0.5  0.5  0.0  0.5  0.5  0.0  0.0 -1.0  0.0  0.5  0.0  0.0  0.0
#> [256] -1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.5  0.5  0.0  0.5 -1.0  0.0  0.5  0.0
#> [271]  0.5  0.5 -1.0  0.5  0.0 -1.0  0.0  0.5 -1.0 -0.5  0.0  0.0 -0.5  0.0  0.0
#> [286]  0.0  0.0  0.0  0.0  0.5  0.0 -1.0  0.5  0.0  0.0  0.0  0.5  0.0  0.0 -0.5
#> [301]  0.0  0.0 -1.0 -1.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.5  0.5
#> [316]  0.5  0.0  0.0 -1.0  0.5  0.5  0.0 -0.5  0.0  0.5  0.0  0.0  0.0  0.0  0.5
#> [331] -1.0 -0.5  0.0  0.0  0.0  0.0  0.0  0.0  0.0 -1.0  0.0  0.0  0.5 -1.0  0.0
#> [346]  0.0 -0.5  0.0  0.0  0.0  0.5 -0.5 -1.0  0.5 -1.0  0.0  0.0  0.0  0.0  0.0
#> [361]  0.0  0.0 -1.0  0.0  0.0  0.0  1.0  0.5  0.0  0.5  0.0  0.0 -1.0  0.0  0.0
#> [376] -1.0 -1.0 -1.0  0.0  0.0  0.0  0.5  0.0  0.0  0.0 -1.0  0.0 -0.5 -0.5  0.5
#> [391]  0.5  0.5  0.0  0.5  0.0  0.5  0.5  0.5  0.0  0.5  0.0 -1.0 -0.5 -1.0 -1.0
#> [406] -1.0  0.0  0.0  0.0 -0.5  0.5  0.0  0.5 -0.5  1.0  0.0  0.5  0.5  0.5  0.0
#> [421]  0.5  0.0  0.0  0.0 -1.0 -1.0  0.0 -1.0  0.5  0.5  0.0  0.0 -0.5  0.0 -0.5
#> [436] -1.0  0.5  0.0  0.0  0.0  0.0 -1.0  0.5 -0.5  1.0  1.0  0.0  0.0  0.5  0.5
#> [451]  0.0  0.0  0.0 -1.0  0.5  0.0  0.0  0.5  0.0 -0.5  0.0  0.0  0.0  0.5  0.5
#> [466]  1.0  0.0 -1.0  0.5  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.5  0.0
#> [481] -0.5  0.5  0.0  0.0  0.0  0.0  0.0 -0.5  0.0  0.0  0.5  0.0  0.5  0.0  0.0
#> [496]  0.0  0.0  0.0  0.0  0.5