Unified Model

Quantifies the chill accumulation by means of converting temperatures to chill-units.

unified_model(x, a = 0.89, b = -28.87, c = -19.44, total = TRUE)

Arguments

x
Vector containing temperature values (Celsius-degree).

a
Parameter defined by the user as the function limits. Default is 0.89.

b
Parameter defined by the user as the function limits. Default is -28.87.

c
Parameter defined by the user as the function limits. Default is -19.44.

total
TRUE Shows the total value of accumulation, FALSE shows the value of chill-unit for each temperature (TRUE is default).

Value

The function returns values the chill-units for each temperature of vector (Total = FALSE), or returns the chill-units accumulation (Total = TRUE).

Details

The coefficients used in this model are adjusted for the apple tree.The model is based on chill-units, but the limits are unknown.

References

Chuine, I. et al. 2016. Can phenological models predict tree phenology accurately in the future? The unrevealed hurdle of endodormancy break. Global Change Biology.

Chuine, Isabelle. 2000. A unified model for budburst of trees. Journal of Theoretical Biology

Examples


x <- rnorm(500,10,3)
unified_model(x)

#> [1] 424.9509

unified_model(x, total = FALSE)

#>   [1] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>   [6] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [11] 1.013011e-47 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [16] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.115708e-10
#>  [21] 1.000000e+00 1.000000e+00 1.000000e+00 9.999577e-01 1.000000e+00
#>  [26] 1.000000e+00 2.562459e-03 1.000000e+00 1.000000e+00 1.000000e+00
#>  [31] 1.000000e+00 1.000000e+00 9.999976e-01 1.000000e+00 9.993803e-01
#>  [36] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [41] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [46] 1.000000e+00 9.039594e-16 1.000000e+00 9.999192e-01 1.000000e+00
#>  [51] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [56] 1.000000e+00 1.000000e+00 1.000000e+00 9.999893e-01 1.000000e+00
#>  [61] 1.000000e+00 5.404000e-49 1.000000e+00 1.000000e+00 4.199548e-06
#>  [66] 1.341508e-42 1.000000e+00 1.197993e-15 1.000000e+00 1.000000e+00
#>  [71] 1.023747e-23 1.000000e+00 1.140522e-04 1.000000e+00 1.000000e+00
#>  [76] 1.000000e+00 8.085723e-07 1.000000e+00 1.000000e+00 2.207541e-11
#>  [81] 7.155704e-48 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [86] 1.000000e+00 1.000000e+00 2.491622e-22 1.000000e+00 7.989612e-16
#>  [91] 1.000000e+00 1.000000e+00 9.999049e-01 1.000000e+00 1.942473e-64
#>  [96] 1.000000e+00 1.000000e+00 1.485574e-22 1.905440e-40 1.000000e+00
#> [101] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 9.999824e-01
#> [106] 1.000000e+00 6.476497e-35 1.000000e+00 1.000000e+00 1.000000e+00
#> [111] 9.415052e-16 9.999995e-01 1.141739e-21 1.000000e+00 1.000000e+00
#> [116] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [121] 1.000000e+00 1.000000e+00 1.000000e+00 3.405178e-06 1.000000e+00
#> [126] 1.000000e+00 1.000000e+00 9.999733e-01 9.999994e-01 9.999966e-01
#> [131] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [136] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [141] 1.000000e+00 1.000000e+00 9.999985e-01 1.000000e+00 1.000000e+00
#> [146] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [151] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 5.536032e-04
#> [156] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [161] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [166] 1.000000e+00 1.000000e+00 1.000000e+00 8.949073e-29 1.000000e+00
#> [171] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [176] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 4.818309e-24
#> [181] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [186] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 2.295510e-05
#> [191] 9.895616e-01 1.000000e+00 1.000000e+00 1.000000e+00 3.005181e-11
#> [196] 1.000000e+00 7.325434e-01 1.000000e+00 1.841105e-07 1.892147e-07
#> [201] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [206] 9.999895e-01 1.000000e+00 1.000000e+00 9.484936e-05 1.000000e+00
#> [211] 1.000000e+00 3.552207e-13 9.526477e-01 1.000000e+00 1.000000e+00
#> [216] 1.000000e+00 1.000000e+00 1.000000e+00 9.999883e-01 1.000000e+00
#> [221] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [226] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [231] 3.423683e-95 1.000000e+00 8.339399e-56 9.980367e-01 1.000000e+00
#> [236] 9.995051e-01 4.411163e-11 1.000000e+00 1.257125e-04 1.000000e+00
#> [241] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 3.337905e-12
#> [246] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [251] 1.000000e+00 1.000000e+00 8.434215e-40 1.000000e+00 1.000000e+00
#> [256] 9.997149e-01 1.294244e-02 1.000000e+00 1.299978e-86 1.000000e+00
#> [261] 1.000000e+00 1.000000e+00 6.374178e-07 1.000000e+00 1.000000e+00
#> [266] 3.787478e-42 2.807167e-10 1.000000e+00 1.000000e+00 1.000000e+00
#> [271] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [276] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [281] 1.000000e+00 1.000000e+00 1.000000e+00 1.446668e-11 1.845988e-04
#> [286] 1.000000e+00 1.000000e+00 2.916914e-01 1.179649e-27 1.000000e+00
#> [291] 1.000000e+00 4.115822e-02 1.000000e+00 1.000000e+00 1.000000e+00
#> [296] 1.000000e+00 4.091471e-04 1.000000e+00 1.970164e-08 1.000000e+00
#> [301] 1.000000e+00 7.800437e-30 1.000000e+00 9.902800e-01 1.000000e+00
#> [306] 1.000000e+00 1.000000e+00 8.054556e-11 1.000000e+00 1.000000e+00
#> [311] 1.000000e+00 1.000000e+00 5.079561e-45 1.000000e+00 1.000000e+00
#> [316] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [321] 1.000000e+00 2.879131e-30 1.000000e+00 1.000000e+00 1.000000e+00
#> [326] 1.000000e+00 9.996114e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> [331] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [336] 1.000000e+00 1.000000e+00 3.998486e-29 1.000000e+00 1.000000e+00
#> [341] 6.233312e-04 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [346] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [351] 1.000000e+00 6.457560e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> [356] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [361] 1.000000e+00 4.890354e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> [366] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [371] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 8.709868e-17
#> [376] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [381] 4.084845e-04 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [386] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [391] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [396] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [401] 4.536513e-20 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [406] 1.011890e-46 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [411] 1.000000e+00 5.984844e-16 1.000000e+00 1.000000e+00 1.000000e+00
#> [416] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [421] 9.933099e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [426] 1.000000e+00 1.000000e+00 2.628938e-19 8.202147e-01 1.000000e+00
#> [431] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.104113e-50
#> [436] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [441] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [446] 9.999977e-01 1.000000e+00 1.000000e+00 1.000000e+00 3.380913e-02
#> [451] 1.000000e+00 1.000000e+00 5.784143e-04 1.000000e+00 9.583357e-01
#> [456] 1.000000e+00 8.093667e-12 1.000000e+00 1.000000e+00 3.526130e-44
#> [461] 1.000000e+00 1.231679e-02 1.000000e+00 1.000000e+00 9.999994e-01
#> [466] 1.000000e+00 7.087489e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> [471] 1.000000e+00 1.000000e+00 6.387556e-05 1.000000e+00 1.000000e+00
#> [476] 4.029115e-03 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [481] 1.000000e+00 1.391817e-15 1.000000e+00 1.000000e+00 1.000000e+00
#> [486] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 2.907485e-01
#> [491] 1.000000e+00 1.000000e+00 9.997838e-01 1.559111e-45 2.675550e-17
#> [496] 1.975698e-13 9.999993e-01 9.823357e-01 1.000000e+00 1.000000e+00